// client/models/UseServiceDto.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type UseServiceDto = {
    /**
     * QR Code to scan
     */
    qrCode: string;
    /**
     * Service ID to use
     */
    serviceId: string;
    /**
     * User ID who is performing the service
     */
    usedById?: string;
    /**
     * Notes about the service usage
     */
    notes?: string;
};